pyproject.toml
[tool.poetry]
name = "minigis"
version = "0.1.0"
description = ""
authors = ["ArthurGolubev <gg.arthur.go@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.10"
fastapi = {extras = ["all"], version = "^0.85.1"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.137.0"}
earthengine-api = "^0.1.328"
loguru = "^0.6.0"
google-cloud-storage = "^2.5.0"
geopandas = "^0.12.1"
rasterio = "^1.3.3"
xarray = "^2022.11.0"
rioxarray = "^0.13.1"
earthpy = "^0.9.4"
spectral = "^0.23.1"
scikit-learn = "^1.1.3"
cvxopt = "^1.3.0"
pyproj = "^3.4.0"
pyshp = "^2.3.1"
yadisk = "^1.2.17"
sqlmodel = "^0.0.8"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyhumps = "^3.8.0"
psycopg2-binary = "^2.9.5"
websockets = "^11.0.1"
python-socketio = {extras = ["client"], version = "^5.8.0"}
alembic = "^1.10.3"
google-cloud-bigquery = "^3.10.0"
ruff = "^0.0.287"
pytest = "^7.4.2"
httpx = "^0.25.0"
pytest-mock = "^3.11.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
